ASSERTION FAILED: containingBlock.isInFlowPositioned() while running `the-anchor-attribute-003-crash.tentative.html` WPT Test https://bugs.webkit.org/show_bug.cgi?id=267993
Reviewed by NOBODY (OOPS!).

According to the spec [1], some properties, such as "position",
"transform", "will-change", "contain", can cause a box to establish an
absolute positioning containing block and become a containing block for
an absolute positioned box.

It makes this assertion invalid in some cases. We should remove it.

The similar change was done in Blink in [2] where this test was
introduced.

1. https://drafts.csswg.org/css-position/#absolute-positioning-containing-block
2. https://bugs.chromium.org/p/chromium/issues/detail?id=1486148

* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:
* Source/WebCore/rendering/RenderBox.cpp:
(WebCore::RenderBox::containingBlockLogicalHeightForPositioned const):