[GTK][WPE] Detect gamepad button 16 https://bugs.webkit.org/show_bug.cgi?id=268653
Reviewed by Michael Catanzaro.

The Gamepad spec lists 17 buttons, starting at zero. The button number
16 is defined as "Center button in center cluster" [1], e.g. the PS
button in a DualShock 4 or 5, or similar ones in other controllers.

The libmanette gamepad backend only lists up to button number 15, which
was what was under the spec when the backend was introduced.

Add the button 16 to the listing as well. Listing is all that is needed
since the rest of the code already detects and maps this button as
expected

[1] https://www.w3.org/TR/gamepad/#dfn-standard-gamepad

* Source/WebCore/platform/gamepad/manette/ManetteGamepad.cpp:
(WebCore::toStandardGamepadButton):
* Source/WebCore/platform/gamepad/manette/ManetteGamepad.h:

Canonical link: https://commits.webkit.org/274022@main