[Site Isolation] Text input fails with `shouldSuppressTextInputFromEditing` https://bugs.webkit.org/show_bug.cgi?id=268674 rdar://120437644
Reviewed by NOBODY (OOPS!).

`shouldSuppressTextInputFromEditing` requires to check if the main frame has a provisional load state,
which is not possible when the main frame is being hosted in another process. This security feature is
not relevant for a cross-origin iframe, so we should just skip it when the main frame is remote.

* Source/WebCore/dom/EventDispatcher.cpp:
(WebCore::shouldSuppressEventDispatchInDOM):
* Source/WebCore/editing/Editor.cpp:
(WebCore::Editor::canPaste const):
(WebCore::Editor::shouldInsertText const):
* Source/WebCore/editing/EditorCommand.cpp:
(WebCore::allowExecutionWhenDisabledPaste):