[threaded-animation-resolution] ensure we stop animation-related tasks when an animated node is removed https://bugs.webkit.org/show_bug.cgi?id=268717 rdar://122278779
Reviewed by Simon Fraser.

It is possible a `RemoteLayerTreeNode` will be removed while it is registered with
`RemoteScrollingCoordinatorProxyIOS` as an animated node in `m_animatedNodeLayerIDs`.
We need to ensure than when a node is removed from the remote layer tree, any
notion of it being animated is cleared by calling `animationsWereRemovedFromNode()`
in `RemoteLayerTreeHost::layerWillBeRemoved()`.

* Source/WebKit/UIProcess/RemoteLayerTree/RemoteLayerTreeHost.mm:
(WebKit::RemoteLayerTreeHost::layerWillBeRemoved):

Canonical link: https://commits.webkit.org/274086@main