[GStreamer] BubblewrapLauncher sandbox lacks directory for the gstreamer user registry cache directory https://bugs.webkit.org/show_bug.cgi?id=268759
Reviewed by Michael Catanzaro.

Grant read-write access in the default GStreamer registry path within the WebProcess sandbox. This
is safe because the GStreamer registry file format is binary and the file is not executable. The
internal GStreamer code responsible for loading this file is able to handle parsing errors.

For additional context, the registry file stores informations about the plugins available on the
system, so that the plugin scanner doesn't need to rescan plugins every time GStreamer is
initialized.

* Source/WebKit/UIProcess/Launcher/glib/BubblewrapLauncher.cpp:
(WebKit::bindGStreamerData):

Canonical link: https://commits.webkit.org/274144@main