Text inserted after a <picture> element ends up inside the element https://bugs.webkit.org/show_bug.cgi?id=268895 rdar://122449003
Reviewed by NOBODY (OOPS!).

When attempting to insert text (or any node) after a <picture> element, position
canonicalization results in insertion occuring inside the <picture> element. This
is undesirable, since <picture> elements are only intended to contain an <img> and
one or more <source> elements.

* LayoutTests/editing/inserting/insert-text-after-picture-expected.txt: Added.
* LayoutTests/editing/inserting/insert-text-after-picture.html: Added.
* Source/WebCore/html/HTMLPictureElement.h:

`canContainRangeEndPoint` is used by `canHaveChildrenForEditing`. Ensure that
<picture> cannot gain children from editing, similar to other elements that
behave as a single unit.