[WebGPU] ShaderModule::createLibrary may produce invalid metal with custom layouts https://bugs.webkit.org/show_bug.cgi?id=268898 <radar://122256850>
Reviewed by NOBODY (OOPS!).

This is accounting for over 50% of fuzzing crashes, it is because a custom
pipeline layout will impact metal source generation and the combination
may not be valid. So we should only assert in fuzzer builds when auto layout
is used.

Otherwise we return an invalid shader module, which is correct because an
incompatible pipeline layout was used.

* Source/WebGPU/WebGPU/Pipeline.mm:
(WebGPU::createLibrary):
* Source/WebGPU/WebGPU/ShaderModule.h:
* Source/WebGPU/WebGPU/ShaderModule.mm:
(WebGPU::ShaderModule::createLibrary):
(WebGPU::earlyCompileShaderModule):