Fix sandbox compilation error related to empty parameters https://bugs.webkit.org/show_bug.cgi?id=268916 rdar://88915314
Reviewed by David Kilzer and Brent Fulgham.

On rare occasions, see https://bugs.webkit.org/show_bug.cgi?id=222233 for more details, the sandbox string
parameters provided to the Networking process' sandbox can be empty. This will cause the compilation of the
sandbox to fail, and the Networking process will crash on startup being unable to apply the sandbox. We can
guard against this in the sandbox by checking the string length of the parameters. We already do this for
the GPU process and WebContent process.

* Source/WebKit/NetworkProcess/mac/com.apple.WebKit.NetworkProcess.sb.in:

Canonical link: https://commits.webkit.org/274228@main