[UnifiedPDF] Selections should be extensible on shift-click https://bugs.webkit.org/show_bug.cgi?id=268617 rdar://122179325
Reviewed by Tim Horton.

This patch introduces support for extending selections by click/drag
operations while the shift key is active.

When beginning a selection, we check if the shift key is pressed, and
generate an extended selection from the start of the existing selection
(if any) to the mouse click position. We adopt the
-[PDFSelection firstCharCenter] SPI to obtain the start point of our
extended selection.

Other than that, we fall back to existing code which makes sure we
extend our current selection if we're doing a drag operation.

* Source/WebKit/Platform/spi/Cocoa/PDFKitSPI.h:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.h:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.mm:
(WebKit::UnifiedPDFPlugin::extendCurrentSelectionIfNeeded):
(WebKit::UnifiedPDFPlugin::beginTrackingSelection):

Canonical link: https://commits.webkit.org/274262@main