[Site Isolation] Exposed content rect of subframe view is wrong https://bugs.webkit.org/show_bug.cgi?id=268647 rdar://122199122
Reviewed by Simon Fraser and Alex Christensen.

On iOS, exposedContentRect of FrameView is used to decide visible area that needs compositing update (see
RenderLayerCompositor::visibleRectForLayerFlushing and RenderLayerCompositor::flushPendingLayerChanges). Currently
exposedContentRect is set via DrawingArea (and each DrawingArea has one content rect), and it is only set on the main
frame view (as child frames can inheirt the value from their parent, see ScrollView::exposedContentRect()). That means
frames in the subframe process do not have the correct content rect and thus do not render. To fix that, now we set the
exposedContentRect of those frames to be the frame rect. This is a temporary fix that enables us to run tests; we are
going to make a complete fix that ensures UI process sends geometry updates to web process in rdar://122429810.

* LayoutTests/platform/ios/TestExpectations:
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::updateFrameSize):

Canonical link: https://commits.webkit.org/274302@main